A lot of improvements brought over from an in progress game. Best way to build a game framework is to build it out.

+++ Assets/Scripts/Extensions/GameObjects/Rpc/Serialization.cs
@@ -82,7 +82,7 @@ namespace Extensions.GameObjects.Rpc
                     writer.WriteValueSafe(longField);
                     break;
                 case ulong ulongField:
-                    writer.WriteValueSafe(SerializationTypes.UInt);
+                    writer.WriteValueSafe(SerializationTypes.ULong);
                     writer.WriteValueSafe(ulongField);
                     break;
                 case bool boolField:
